+++ src/main/java/com/jfcm/manda/bookingmanagerapi/model/RoomStatusEnum.java
@@ -1,0 +1,17 @@
+/*
+ *  RoomStatusEnum.java
+ *
+ *  Copyright © 2023 ING Group. All rights reserved.
+ *
+ *  This software is the confidential and proprietary information of
+ *  ING Group ("Confidential Information").
+ */
+package com.jfcm.manda.bookingmanagerapi.model;
+
+public enum RoomStatusEnum {
+
+  available,
+  reserved,
+  taken,
+  not_define
+}