+++ src/main/java/com/jfcm/manda/bookingmanagerapi/service/AvailabilityCalendarService.java
@@ -2,7 +2,8 @@ package com.jfcm.manda.bookingmanagerapi.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 
-@FunctionalInterface
-public interface PersistYearlyDates {
+public interface AvailabilityCalendarService {
   void generateYearDates() throws JsonProcessingException;
+
+  void deletePreviousYearDates() throws JsonProcessingException;
 }