+++ src/test/java/com/jfcm/manda/bookingmanagerapi/resource/ClustersResourcesTest.java
@@ -1,3 +1,10 @@
+/**
+ * {@link com.jfcm.manda.bookingmanagerapi.resource.ClustersResourcesTest}.java
+ * Copyright © 2023 JFCM. All rights reserved. This software is the confidential and
+ * proprietary information of JFCM Mandaluyong
+ *
+ * @author Ronald Cando
+ */
 package com.jfcm.manda.bookingmanagerapi.resource;
 
 import static org.hamcrest.Matchers.is;