+++ src/test/java/com/jfcm/manda/bookingmanagerapi/utils/TestUtils.java
@@ -1,14 +1,18 @@
-/*
- *  TestUtils.java
+/**
+ * {@link com.jfcm.manda.bookingmanagerapi.utils.TestUtils}.java
+ * Copyright © 2023 JFCM. All rights reserved. This software is the confidential and
+ * proprietary information of JFCM Mandaluyong
  *
- *  Copyright © 2023 ING Group. All rights reserved.
- *
- *  This software is the confidential and proprietary information of
- *  ING Group ("Confidential Information").
+ * @author Ronald Cando
  */
 package com.jfcm.manda.bookingmanagerapi.utils;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import java.io.BufferedReader;
+import java.io.File;
+import java.io.FileNotFoundException;
+import java.io.FileReader;
+import java.io.FileWriter;
 import java.io.IOException;
 import java.io.InputStream;
 
@@ -25,5 +29,4 @@ public class TestUtils {
       throw new RuntimeException(e);
     }
   }
-
 }