+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.4] (2023-11-07)
+- Added logic for handling reservation in availability_calendar table
+  - Added handler for pass date
+  - Added handler for checking if room is available or not on a given date
+  - Added AvailableRoomOnDateRepositoryImpl, AvailableDateRepository, AvailableRoomOnDateCustom, AvailableRoomOnDateRepository
+
 ## [1.0.3] (2023-11-06)
 - Added JavaDoc as description to some classes
 - Clean-up code